Do not checkout!  Broken/partial commit. I'm too tired to even finish this enough to compile let alone do the tests.  I know this is what git stash is for but I'm too lazy to look it up and use it now.
Changes include adding test stuff to main.c for templates
and vector of vectors, realizing a bunch of issues when I
went to write the vector of vector tests, adding stack
init/constructor functions, need to add stack frees.

Need to read/edit the comments and add some to the new functions.
There need to be wrappers around strdup and free for vector
to imitate vector_s since vector calls elem_free on the address
of the element not the element itself (in vector_s the element is
the pointer and free is called on it directly).  Likewise, elem_init
takes 2 args, src and dest etc. so strdup would be called inside.

The reason I can't change it to call elem_free directly is it's
a byte array and it would covert the first byte to an address not
the full element (and the ways around this, knowing the size of a
pointer etc., are ugly and would subtract functionality shown in
the f_struct examples.