I've made a lot of the changes listed in the last commit message. The code currently compiles and passes the tests but the tests vector of vectors and template tests need to be expanded.
I also need to add the pointer type vector to vector_template and add
that to the tests.

I've written a simple python script to generate a .c/.h pair of files
for a given type from vector_template.c/h but you could just do a
search/replace in your favorite editor and save as.  I've included
the vector_short files as examples of what's generated and how to plug
them into the tests.

Still need to implement a few ideas and create a branch or 2 to for different
design/use ideas.