Added cvector.h an all in one file.
I really like the stb_* library (nothings.org) style of having everything
in one file so it's easy to add to a project and doesn't clutter
up the project with source files.  So that's what cvector.h is.

I'll clean it up a little later but I've added a project to premake
to make sure it compiles and run the same tests on it for sanity.

I've renamed a couple of the allocator macros for consistency that
I missed last commit.