Convert to Clean C (aka C++ compatible)
I realize that with extern C blocks in the headers a C++ project
can link with files compiled as C (with all the C99 goodies that C++ doesn't
allow), but that won't work for an all in one header like cvector.h if
it's included in a C++ file.  I suppose even there you could just create
a C file wrapper for the cvector implementation and just have the header
in any C++ file. Like so:

cvectorwrapper.c:

anycppfile.cpp

whatever you want here

In that case even cvector.h wouldn't need to be clean C
just the header extern C wrapped.  At that point though,
you've lost the ease/convenience of just dropping in the all-in-one
header.

Whatever it's all moot because I've made the whole project C++
compatible so it's the intersection of C++ and C89.