Added selection macros to cvector.h
So now if you only use 1 of the 4 vectors you can do this

so you won't be compiling in the other 3 and bloating your binary.
Granted you're saving about 10 KB which is miniscule these days but
still, it's worth having.

I also removed extraneous blank lines that always bothered me.