Move config and selection macros into the header
This fixes the problem where you use the template macros in a different
translation unit than the CVECTOR_IMPLEMENTATION.  Before, when the
memory, memmove and assert macros were set only in the IMPLEMENTATION,
they would of course be undefined in the template macro definition expansion in
another TU.

As for why vector selection macros were outside of both the header in
the first place, I have no idea.  Probably because they weren't part of
any individual file so I was thinking of them as separate when I wrote
the script but knew they needed to be at the top.